+++ rivercharts/views.py
@@ -5,6 +5,7 @@ import os
 import requests
 import math
 import pytz
+from datetime import datetime
 import json
 import pandas as pd
 import plotly.graph_objs as go
@@ -14,7 +15,7 @@ from datetime import datetime, timedelta
 from django.shortcuts import render
 from django.shortcuts import redirect
 from django.http import JsonResponse
-from .config import USE_DUMMY_DATA, USE_DEMO_DATA, SITE_CODE, PARAMETER_CODE, START_DATE, API_URL, STATION_NAME, REQUEST_TIMEOUT_SECONDS
+from .config import USE_DUMMY_DATA, USE_DEMO_DATA, SITE_CODE, PARAMETER_CODE, START_DATE, END_DATE, TIME_ZONE, API_URL, STATION_NAME, REQUEST_TIMEOUT_SECONDS
 
 # These are the config variables from config.py
 config_use_dummy_data = USE_DUMMY_DATA
@@ -22,13 +23,15 @@ config_use_demo_data = USE_DEMO_DATA
 config_site_code = SITE_CODE
 config_parameter_code = PARAMETER_CODE
 config_start_date = START_DATE
+config_end_date = END_DATE
+config_time_zone = TIME_ZONE
 config_api_url = API_URL
 config_station_name = STATION_NAME
 requests_timeout_seconds = REQUEST_TIMEOUT_SECONDS
 
 # These are the default values
 current_date = datetime.now().strftime("%Y-%m-%d")
-start_date = datetime(2015, 7, 1, tzinfo=pytz.utc)
+start_date = datetime(2015, 7, 1, tzinfo=pytz.utc) # Start date for dummy data
 dummy_data = {
     'value': {
         'timeSeries': [{
@@ -36,7 +39,7 @@ dummy_data = {
                 'value': [{
                     'dateTime': (start_date + timedelta(days=i)).strftime("%Y-%m-%dT%H:%M:%S.%f%z"),
                     'value': str(10 + 5*math.sin(i/10))
-                } for i in range(3000)]  # 2000 or however many entries you want
+                } for i in range(3000)]  # 3000 or however many entries you want (days)
             }]
         }]
     }
@@ -60,7 +63,7 @@ def river_graph_data(request):
     api_url = config_api_url
     site_code = config_site_code
     start_date = config_start_date
-    end_date = current_date
+    end_date = current_date if config_end_date == "" else config_end_date
     parameter_code = config_parameter_code
 
     params = {
@@ -111,17 +114,37 @@ def river_graph_data(request):
     time_series = data['value']['timeSeries'][0]
     values = time_series['values'][0]['value']
 
+    # Time zone for EST/EDT
+    time_zone = pytz.timezone(config_time_zone)
+
     # Create a list of dictionaries to store the data
     data_list = []
     for entry in values:
+        # date_time = datetime.strptime(entry['dateTime'], "%Y-%m-%dT%H:%M:%S.%f%z").strftime('%B %d, %Y %I:%M:%S %p EST')
+        # Step 1: Parse the datetime with the time zone information
+        parsed_date_time = datetime.strptime(entry['dateTime'], "%Y-%m-%dT%H:%M:%S.%f%z")
+        # Step 2: Convert to the desired time zone
+        date_time = parsed_date_time.astimezone(time_zone)
+        # Step 3: Format the datetime for display
+        formatted_date_time = date_time.strftime('%B %d, %Y %I:%M:%S %p %Z')
+        
         data_list.append({
             'Date Time': datetime.strptime(entry['dateTime'], "%Y-%m-%dT%H:%M:%S.%f%z"),
             'River Level': float(entry['value']),
             'Date': datetime.strptime(entry['dateTime'], "%Y-%m-%dT%H:%M:%S.%f%z").strftime('%B %d, %Y %I:%M:%S %p EST')
         })
+        last_date_time = formatted_date_time  # Update last_date_time to the most recent entry
 
     # Create a dataframe from the list of dictionaries
     df = pd.DataFrame(data_list)
+
+    # Debug: Print the last few rows of the dataframe to ensure it includes the latest date
+    # print("Last few rows of the DataFrame:")
+    # print(df.tail())
+    
+    # Optional: Resample the DataFrame to hourly means vs. the current more granular approach
+    # FYI: There is no code to support this as of now, this would be the start below:
+    # df_resampled = df.resample('H', on='Date Time').mean().reset_index()
 
     # Create line trace
     fig = px.line(df, x='Date Time', y='River Level', title='River Levels')
@@ -150,7 +173,15 @@ def river_graph_data(request):
         floated_status = row['Floated']
 
         # Determine the color based on the Floated status
-        color = "red" if floated_status == "N" else "green"
+        if floated_status == "N":
+            color = "red" # Red: #FF0000
+        elif floated_status == "Y":
+            color = "green" # Green: #008000
+        else:
+            color = "orange" # Orange: #FFA500
+
+        # Determine the floated status text
+        floated_status_text = "No" if floated_status == "N" else ("Yes" if floated_status == "Y" else "Skipped")
 
         # Find the closest date in df to the float_date and get its river level
         closest_date = df.iloc[((df['Date Time'].apply(lambda x: x.timestamp() * 1e9) - float_date.value).abs()).argsort()[:1]]
@@ -190,7 +221,7 @@ def river_graph_data(request):
                     name=f"Float {float_number} - {river_level:.2f} ft",  # This line sets the name in the legend
                     text=[str(float_number)],  # This places the float number next to the point
                     hoverinfo="text",
-                    hovertext=[f"Date: {float_date.strftime('%m-%d-%Y')}<br>Float #: {float_number}<br>Height: {river_level:.2f} ft<br>Floated: {'Yes' if floated_status == 'Y' else 'No'}"]
+                    hovertext=[f"Date: {float_date.strftime('%m-%d-%Y')}<br>Float #: {float_number}<br>Height: {river_level:.2f} ft<br>Floated: {floated_status_text}"]
                 )
             )
 
@@ -204,6 +235,9 @@ def river_graph_data(request):
 
             x_min = df['Date Time'].min()  # Get the earliest date
             x_max = df['Date Time'].max()  # Get the latest date
+
+            # Debug: Print the min and max x-axes
+            # print(f"x_min: {x_min}, x_max: {x_max}")
 
             fig.update_layout(
                 font=dict(family='GochiHand, sans-serif'), # This changes the font for the entire plot
@@ -231,4 +265,4 @@ def river_graph_data(request):
     graph_div = fig.to_html(full_html=True)
 
     # Return the plot as JSON
-    return JsonResponse({'graph_div': graph_div, 'station_name': config_station_name})
+    return JsonResponse({'graph_div': graph_div, 'station_name': config_station_name, 'last_date_time': last_date_time})